Single OEP and EMCTS
Single OEP and EMCTS with params for node size
