+++ play.json
@@ -7,7 +7,7 @@
   "Search Depth" : 5,
   "Force End" : false,
   "Population Size" : 20,
-  "Players" : ["OEPS", "EMCTSS", "OEP", "EMCTS"],
+  "Players" : ["OEP", "EMCTS", "OEP", "EMCTS"],
   "Tribes" : ["Xin Xi", "Imperius", "Bardur", "Oumaji"],
   "Verbose" : false,
