--- src/players/EMCTSS/EMCTSSParams.java
@@ -1,8 +1,0 @@
-package players.EMCTSS;
-
-import players.heuristics.AlgParams;
-
-public class EMCTSSParams extends AlgParams {
-    public float bias = 1f;
-    public int depth = 10;
-}
