--- src/players/OEPS/OEPSParams.java
@@ -1,9 +1,0 @@
-package players.OEPS;
-import players.heuristics.AlgParams;
-
-public class OEPSParams extends AlgParams {
-
-    public int POP_SIZE = 50;
-    public double MUTATION_RATE = 0.1;
-    public double KILL_RATE = 0.5;
-}
