+++ src/players/emcts/EMCTSAgent.java
@@ -50,9 +50,9 @@ public class EMCTSAgent extends Agent {
         int remainingLimit = 5;
         boolean stop = false;
 
-        if(this.returnAction){
+        if (this.returnAction) {
             Action action;
-            if(bestNode.getSequence().size() == 1){
+            if (bestNode.getSequence().size() == 1) {
                 returnAction = false;
             }
             action = bestNode.returnNext();
@@ -66,7 +66,7 @@ public class EMCTSAgent extends Agent {
 
         bestNode = root;
         fmCallsRun = 0;
-        while(!stop) {
+        while (!stop) {
             numIters++;
             ElapsedCpuTimer elapsedTimerIteration = new ElapsedCpuTimer();
             fmCallsRun = 0;
@@ -74,21 +74,21 @@ public class EMCTSAgent extends Agent {
             EMCTSTreeNode toMutate = nodeToExpand();
 
             int depth = 0;
-            while(depth < params.depth){
+            while (depth < params.depth) {
                 EMCTSTreeNode child = mutate(toMutate, gs.copy());
                 eval(gs.copy(), child);
 
-                if(child.getValue() > bestNode.getValue()){
+                if (child.getValue() > bestNode.getValue()) {
                     bestNode = child;
                 }
 
                 child.visited();
-                extend(child, gs.copy());
+                //extend(child, gs.copy());
                 toMutate = child;
                 depth++;
             }
 
-            for(int i = 0; i < params.depth; i++){
+            for (int i = 0; i < params.depth; i++) {
                 toMutate.refreshScore(params.bias);
                 toMutate = toMutate.getParent();
             }
@@ -99,20 +99,26 @@ public class EMCTSAgent extends Agent {
                 avgTimeTaken = acumTimeTaken / numIters;
                 remaining = ect.remainingTimeMillis();
                 stop = remaining <= 2 * avgTimeTaken || remaining <= remainingLimit;
-                if(stop){this.returnAction = true;}
+                if (stop) {
+                    this.returnAction = true;
+                }
             } else if (params.stop_type == params.STOP_ITERATIONS) {
                 stop = numIters >= params.num_iterations;
-                if(stop){this.returnAction = true;}
+                if (stop) {
+                    this.returnAction = true;
+                }
             } else if (params.stop_type == params.STOP_FMCALLS) {
                 stop = (fmCallsCount > params.num_fmcalls) || (fmCallsRun > (params.num_fmcalls - fmCallsCount));
-                if(stop){this.returnAction = true;}
+                if (stop) {
+                    this.returnAction = true;
+                }
             }
         }
-        //System.out.println(fmCallsCount);
+        System.out.println(fmCallsCount);
 
         Action action = null;
-        if(this.returnAction){
-            if(bestNode.getSequence().size() == 1){
+        if (this.returnAction) {
+            if (bestNode.getSequence().size() == 1) {
                 returnAction = false;
             }
             action = bestNode.returnNext();
@@ -126,39 +132,39 @@ public class EMCTSAgent extends Agent {
         return null;
     }
 
-    public void eval(GameState gs, EMCTSTreeNode node){
-        for(Action move : node.getSequence()){
+    public void eval(GameState gs, EMCTSTreeNode node) {
+        for (Action move : node.getSequence()) {
             advance(gs, move);
         }
         node.setGs(gs);
         node.setValue(heuristic.evaluateState(gs));
     }
 
-    private EMCTSTreeNode nodeToExpand(){
+    private EMCTSTreeNode nodeToExpand() {
         ArrayList<EMCTSTreeNode> children = root.getChildren();
 
         EMCTSTreeNode toMutate = null;
 
-        if(children.size() == 0){
+        if (children.size() == 0) {
             toMutate = root;
-        }else{
+        } else {
             boolean found = false;
             double nodeScore = 0;
             EMCTSTreeNode nodeOn = root;
             nodeScore = nodeOn.getScore();
-            while(!found){
-                if(nodeOn.getChildren().size() == 0){
+            while (!found) {
+                if (nodeOn.getChildren().size() == 0) {
                     found = true;
                     toMutate = nodeOn;
-                }else{
+                } else {
                     boolean bigger = false;
-                    for(EMCTSTreeNode node : nodeOn.getChildren()){
-                        if(node.getScore() > nodeScore){
+                    for (EMCTSTreeNode node : nodeOn.getChildren()) {
+                        if (node.getScore() > nodeScore) {
                             bigger = true;
                             nodeOn = node;
                             nodeScore = node.getScore();
                         }
-                        if(!bigger){
+                        if (!bigger) {
                             found = true;
                             toMutate = nodeOn;
                         }
@@ -170,53 +176,64 @@ public class EMCTSAgent extends Agent {
     }
 
     //this version of random Actions will return a list of random actions and tries to make the length as long as possible
-    private ArrayList<Action> randomActions(GameState gs){
+    private ArrayList<Action> randomActions(GameState gs) {
         ArrayList<Action> individual = new ArrayList<>();
-        while (!gs.isGameOver() && (gs.getActiveTribeID() == getPlayerID())){
+        while ((!gs.isGameOver() && (gs.getActiveTribeID() == getPlayerID())) && (individual.size() < (params.NODE_SIZE-1))){
             ArrayList<Action> allAvailableActions = this.allGoodActions(gs, m_rnd);
             Action a = allAvailableActions.get(m_rnd.nextInt(allAvailableActions.size()));
-            if(!(a.getActionType() == Types.ACTION.END_TURN) || allAvailableActions.size() == 1){
+            if (!(a.getActionType() == Types.ACTION.END_TURN) || allAvailableActions.size() == 1) {
                 advance(gs, a);
                 individual.add(a);
             }
         }
+        if(individual.size() == (params.NODE_SIZE-1) && (individual.get((params.NODE_SIZE-2)).getActionType() != Types.ACTION.END_TURN)){
+            ArrayList<Action> allAvailableActions = this.allGoodActions(gs, m_rnd);
+            Action end = null; ;
+            for(Action a : allAvailableActions){
+                if(a.getActionType() == Types.ACTION.END_TURN){
+                    end = a;
+                    break;
+                }
+            }
+            individual.add(end);
+        }
         return individual;
     }
 
-    private void advance(GameState gs, Action move){
+    private void advance(GameState gs, Action move) {
         this.fmCallsCount++;
         this.fmCallsRun++;
-        gs.advance(move,true);
+        gs.advance(move, true);
     }
 
-    private EMCTSTreeNode mutate(EMCTSTreeNode node, GameState gs){
+    private EMCTSTreeNode mutate(EMCTSTreeNode node, GameState gs) {
         GameState clone = gs.copy();
         ArrayList<Action> seq = node.getSequence();
         int moveToMutate = m_rnd.nextInt(seq.size());
         ArrayList<Action> newSeq = new ArrayList<>();
-        for(int i = 0; i < moveToMutate; i++){
-            advance(gs,seq.get(i));
+        for (int i = 0; i < moveToMutate; i++) {
+            advance(gs, seq.get(i));
             newSeq.add(seq.get(i));
         }
         ArrayList<Action> allAvailableActions = this.allGoodActions(gs, m_rnd);
         Action a = allAvailableActions.get(m_rnd.nextInt(allAvailableActions.size()));
         newSeq.add(a);
 
-        for(int i = moveToMutate +1; i < seq.size(); i++){
+        for (int i = moveToMutate + 1; i < seq.size(); i++) {
             newSeq.add(seq.get(i));
         }
 
         newSeq = repair(newSeq, clone);
-        EMCTSTreeNode newNode = new EMCTSTreeNode(newSeq,node);
+        EMCTSTreeNode newNode = new EMCTSTreeNode(newSeq, node);
         node.addChild(newNode);
 
         return newNode;
     }
 
-    private ArrayList<Action> repair(ArrayList<Action> child, GameState gs){
+    private ArrayList<Action> repair(ArrayList<Action> child, GameState gs) {
         ArrayList<Action> repairedChild = new ArrayList<>();
 
-        for(int a = 0 ;a < child.size(); a ++) {
+        for (int a = 0; a < child.size(); a++) {
             if (!(gs.getActiveTribeID() == getPlayerID())) {
                 return repairedChild;
             }
@@ -240,82 +257,82 @@ public class EMCTSAgent extends Agent {
                 repairedChild.add(ac);
             }
         }
-        if(!(repairedChild.get(repairedChild.size()-1).getActionType() == Types.ACTION.END_TURN)){
-            while ((!gs.isGameOver()) && (gs.getActiveTribeID() == getPlayerID())){
-                ArrayList<Action> allAvailableActions = this.allGoodActions(gs, m_rnd);
-                Action a = allAvailableActions.get(m_rnd.nextInt(allAvailableActions.size()));
-                advance(gs, a);
-                repairedChild.add(a);
+        if (!(repairedChild.get(repairedChild.size() - 1).getActionType() == Types.ACTION.END_TURN)) {
+            ArrayList<Action> allAvailableActions = this.allGoodActions(gs, m_rnd);
+            Action end = null; ;
+            for(Action a : allAvailableActions){
+                if(a.getActionType() == Types.ACTION.END_TURN){
+                    end = a;
+                    break;
+                }
             }
+            repairedChild.add(end);
         }
 
         return repairedChild;
     }
 
-    private boolean checkActionFeasibility(Action a, GameState gs)
-    {
-        if(gs.isGameOver())
+    private boolean checkActionFeasibility(Action a, GameState gs) {
+        if (gs.isGameOver())
             return false;
 
-        if(a instanceof UnitAction)
-        {
-            UnitAction ua = (UnitAction)a;
+        if (a instanceof UnitAction) {
+            UnitAction ua = (UnitAction) a;
             int unitId = ua.getUnitId();
             Actor act = gs.getActor(unitId);
-            if(!(act instanceof Unit) || act.getTribeId() != gs.getActiveTribeID())
+            if (!(act instanceof Unit) || act.getTribeId() != gs.getActiveTribeID())
                 return false;
-        }else if (a instanceof CityAction)
-        {
-            CityAction ca = (CityAction)a;
+        } else if (a instanceof CityAction) {
+            CityAction ca = (CityAction) a;
             int cityId = ca.getCityId();
             Actor act = gs.getActor(cityId);
-            if(!(act instanceof City) || act.getTribeId() != gs.getActiveTribeID())
+            if (!(act instanceof City) || act.getTribeId() != gs.getActiveTribeID())
                 return false;
         }
 
         boolean feasible = false;
-        try{
+        try {
             feasible = a.isFeasible(gs);
-        }catch (Exception e) { }
+        } catch (Exception e) {
+        }
 
         return feasible;
     }
 
     // try to see of the node can be extended
-    private void extend(EMCTSTreeNode node, GameState gs){
+    private void extend(EMCTSTreeNode node, GameState gs) {
         GameState clone = gs.copy();
         ArrayList<Action> currentActions = node.getSequence();
         ArrayList<Action> newActions = new ArrayList<>();
-        for(int i = 0; i < currentActions.size()-1; i++){
+        for (int i = 0; i < currentActions.size() - 1; i++) {
             newActions.add(currentActions.get(i));
-            advance(gs,currentActions.get(i));
+            advance(gs, currentActions.get(i));
         }
 
-        if(this.allGoodActions(gs, m_rnd).size() > 1){
-            while (!gs.isGameOver() && (gs.getActiveTribeID() == getPlayerID())){
+        if (this.allGoodActions(gs, m_rnd).size() > 1) {
+            while (!gs.isGameOver() && (gs.getActiveTribeID() == getPlayerID())) {
                 ArrayList<Action> allAvailableActions = this.allGoodActions(gs, m_rnd);
                 Action a = allAvailableActions.get(m_rnd.nextInt(allAvailableActions.size()));
-                if(!(a.getActionType() == Types.ACTION.END_TURN)){
+                if (!(a.getActionType() == Types.ACTION.END_TURN)) {
                     advance(gs, a);
                     newActions.add(a);
-                }else if(allAvailableActions.size() == 1){
+                } else if (allAvailableActions.size() == 1) {
                     advance(gs, a);
                     newActions.add(a);
                     break;
                 }
             }
-        }else{
+        } else {
             return;
         }
-        newActions = repair(newActions,clone.copy());
+        newActions = repair(newActions, clone.copy());
         EMCTSTreeNode newNode = new EMCTSTreeNode(newActions, node.getParent());
         eval(clone, newNode);
-        if(newNode.getValue() > node.getValue()){
+        if (newNode.getValue() > node.getValue()) {
             node.setSequence(newActions);
             node.setValue(newNode.getValue());
         }
         //clear links so that JVM can clear it
         newNode.unlink();
     }
-
 }
