+++ src/players/emcts/EMCTSParams.java
@@ -5,4 +5,5 @@ import players.heuristics.AlgParams;
 public class EMCTSParams extends AlgParams {
     public float bias = 1f;
     public int depth = 10;
+    public int NODE_SIZE = 20;
 }
