+++ src/players/oep/OEPParams.java
@@ -7,4 +7,5 @@ public class OEPParams extends AlgParams {
     public int POP_SIZE = 50;
     public double MUTATION_RATE = 0.1;
     public double KILL_RATE = 0.5;
+    public int NODE_SIZE = 20;
 }
